+++ writer.go
@@ -214,7 +214,7 @@ func (w *Writer) writeQuotedFieldSIMD(field string) error {
 	// Process 32-byte chunks
 	for i+32 <= len(data) {
 		chunk := archsimd.LoadInt8x32((*[32]int8)(unsafe.Pointer(&data[i])))
-		mask := uint32(chunk.Equal(quoteCmp).ToBits())
+		mask := chunk.Equal(quoteCmp).ToBits()
 
 		if mask != 0 {
 			// Found quotes in this chunk - process them
